@@ -24,47 +24,52 @@
             "email": "joe@alt-three.com"
         }
     ],
+    "replace": {
+        "paragonie/random_compat": "*",
+        "symfony/polyfill-mbstring": "*"
+    },
     "require": {
         "php": "^7.1.3",
+        "ext-mbstring": "*",
+        "ext-openssl": "*",
         "ext-xml": "*",
-        "alt-three/badger": "^5.0",
-        "alt-three/bus": "^4.0",
-        "alt-three/emoji": "^6.0",
-        "alt-three/throttle": "^3.0",
-        "alt-three/twitter": "^3.0",
-        "alt-three/validator": "^4.0",
+        "alt-three/badger": "^5.1",
+        "alt-three/bus": "^4.1",
+        "alt-three/emoji": "^7.0",
+        "alt-three/throttle": "^3.1",
+        "alt-three/twitter": "^3.1",
+        "alt-three/validator": "^4.1",
         "aws/aws-sdk-php": "^3.7",
         "barryvdh/laravel-cors": "^0.11.0",
-        "doctrine/dbal": "2.5.13",
+        "doctrine/dbal": "2.9.*",
         "fideloper/proxy": "^4.0",
-        "graham-campbell/binput": "^5.0",
-        "graham-campbell/exceptions": "^11.0",
-        "graham-campbell/markdown": "^10.0",
-        "guzzlehttp/guzzle": "^6.2.1",
-        "jenssegers/date": "^3.2",
-        "laravel/framework": "5.6.*",
+        "graham-campbell/binput": "^6.0",
+        "graham-campbell/exceptions": "^11.1",
+        "graham-campbell/markdown": "^10.2",
+        "guzzlehttp/guzzle": "^6.3.3",
+        "jenssegers/date": "^3.4",
+        "laravel/framework": "5.7.*",
         "laravel/tinker": "^1.0",
-        "laravolt/avatar": "^1.8",
-        "mccool/laravel-auto-presenter": "^7.0",
-        "nexmo/client": "@beta",
+        "laravolt/avatar": "^2.1",
+        "mccool/laravel-auto-presenter": "^7.1",
+        "nexmo/client": "^1.5",
         "pragmarx/google2fa": "^0.7.1",
         "predis/predis": "^1.1",
-        "roumen/feed": "^2.10",
-        "twig/twig": "^1"
+        "twig/twig": "^2.6"
     },
     "require-dev": {
         "ext-sqlite3": "*",
         "alt-three/testbench": "^4.0",
-        "barryvdh/laravel-debugbar": "^3.0",
-        "filp/whoops": "^2.1",
-        "fzaninotto/faker": "^1.6",
-        "graham-campbell/analyzer": "^2.0",
+        "barryvdh/laravel-debugbar": "^3.2",
+        "filp/whoops": "^2.3",
+        "fzaninotto/faker": "^1.8",
+        "graham-campbell/analyzer": "^2.1",
         "graham-campbell/testbench-core": "^3.0",
-        "mockery/mockery": "^1.0",
-        "phpunit/phpunit": "^7.0",
-        "symfony/css-selector": "^3.1",
-        "symfony/dom-crawler": "^3.1",
-        "tightenco/mailthief": "^0.3.2"
+        "mockery/mockery": "^1.2",
+        "phpunit/phpunit": "^7.4",
+        "symfony/css-selector": "^4.1",
+        "symfony/dom-crawler": "^4.1",
+        "tightenco/mailthief": "^0.3.14"
     },
     "suggest": {
         "ext-apc": "APC Support cache driver."
@@ -98,6 +103,9 @@
         ]
     },
     "config": {
+        "platform": {
+            "php": "7.1.3"
+        },
         "preferred-install": "dist",
         "sort-packages": true,
         "optimize-autoloader": true
