@@ -57,6 +57,6 @@ class ComposerServiceProvider extends ServiceProvider
      */
     public function register()
     {
-        //
+        $this->app->singleton(DashboardComposer::class);
     }
 }
