@@ -16,7 +16,6 @@ use CachetHQ\Cachet\Bus\Events\System\SystemWasResetEvent;
 use CachetHQ\Cachet\Bus\Events\System\SystemWasUpdatedEvent;
 use CachetHQ\Cachet\Settings\Cache;
 use Illuminate\Console\Command;
-use Illuminate\Contracts\Config\Repository;
 use Illuminate\Contracts\Events\Dispatcher;
 
 /**
@@ -35,24 +34,15 @@ class CommandSubscriber
     protected $cache;
 
     /**
-     * The config repository instance.
-     *
-     * @var \Illuminate\Contracts\Config\Repository
-     */
-    protected $config;
-
-    /**
      * Create a new command subscriber instance.
      *
-     * @param \CachetHQ\Cachet\Settings\Cache         $cache
-     * @param \Illuminate\Contracts\Config\Repository $config
+     * @param \CachetHQ\Cachet\Settings\Cache $cache
      *
      * @return void
      */
-    public function __construct(Cache $cache, Repository $config)
+    public function __construct(Cache $cache)
     {
         $this->cache = $cache;
-        $this->config = $config;
     }
 
     /**
