@@ -23,13 +23,15 @@ return [
         'group' => [
             'other' => 'その他のコンポーネント',
         ],
+        'select_all'   => '全て選択',
+        'deselect_all' => '全て選択解除',
     ],
 
     // Incidents
     'incidents' => [
         'none'         => 'インシデントはありません',
         'past'         => 'インシデント履歴',
-        'stickied'     => 'Stickied Incidents',
+        'stickied'     => '固定している障害情報',
         'scheduled'    => '計画メンテナンス',
         'scheduled_at' => ', 予定日時 :timestamp',
         'posted'       => '掲載日時 :timestamp',
@@ -75,14 +77,17 @@ return [
 
     // Subscriber
     'subscriber' => [
-        'subscribe'           => '最新のアップデート情報を購読する',
+        'subscribe'           => 'Subscribe to status changes and incident updates',
         'unsubscribe'         => '購読の解除',
         'button'              => '購読',
         'manage_subscription' => '購読の管理',
         'manage'              => [
-            'no_subscriptions' => '現在、すべてのアップデートを購読しています。',
-            'my_subscriptions' => '現在、以下のアップデートを購読しています。',
-            'manage_at_link'   => '購読の管理',
+            'notifications'       => 'Notifications',
+            'notifications_for'   => '通知の管理',
+            'no_subscriptions'    => '現在、すべてのアップデートを購読しています。',
+            'update_subscription' => '購読の管理',
+            'my_subscriptions'    => '現在、以下のアップデートを購読しています。',
+            'manage_at_link'      => '購読の管理',
         ],
         'email' => [
             'subscribe'          => 'メールによるアップデート情報の購読',
@@ -122,9 +127,9 @@ return [
     // Meta descriptions
     'meta' => [
         'description' => [
-            'incident'  => 'Details and updates about the :name incident that occurred on :date',
-            'schedule'  => 'Details about the scheduled maintenance period :name starting :startDate',
-            'subscribe' => 'Subscribe to :app in order to receive updates of incidents and scheduled maintenance periods',
+            'incident'  => ':date に発生した:nameの障害情報',
+            'schedule'  => ':startDate に予定している:name のメンテナンスの詳細',
+            'subscribe' => ':app の障害情報とメンテナンス情報を購読する',
             'overview'  => 'Stay up to date with the latest service updates from :app.',
         ],
     ],
