@@ -26,8 +26,7 @@
                     </div>
                     <div class="form-group">
                         <label>{{ trans('forms.user.password') }}</label>
-                        <input type="password" class="form-control password-strength" name="password" value="" placeholder="{{ trans('forms.user.password') }}">
-                        <div class="strengthify-wrapper"></div>
+                        <input type="password" class="form-control" name="password" value="" placeholder="{{ trans('forms.user.password') }}">
                     </div>
                     @if($currentUser->isAdmin)
                     <div class="form-group">
