@@ -2,7 +2,7 @@
 <div class="section-timeline">
     <h1>{{ trans('cachet.incidents.past') }}</h1>
     @foreach($allIncidents as $date => $incidents)
-    @include('partials.incidents', [compact($date), compact($incidents)])
+    @include('partials.incidents', [@compact($date), @compact($incidents)])
     @endforeach
 </div>
 
