@@ -13,8 +13,8 @@
 <div class="row">
     <div class="col-xs-12 col-lg-offset-2 col-lg-8">
         <div class="text-center margin-bottom">
-            <h1>{{ $appName }} Notifications</h1>
-            <p>Manage notifications for <strong>{{ $subscriber->email }}</strong></p>
+            <h1>{{ $appName }} {{ trans('cachet.subscriber.manage.notifications') }}</h1>
+            <p>{{ trans('cachet.subscriber.manage.notifications_for') }} <strong>{{ $subscriber->email }}</strong></p>
         </div>
         <form action="{{ cachet_route('subscribe.manage', [$subscriber->verify_code], 'post') }}" method="post">
             <input type="hidden" name="_token" value="{{ csrf_token() }}">
@@ -26,9 +26,9 @@
                     <i class="{{ $componentGroup->collapse_class_with_subscriptions($subscriptions) }} group-toggle"></i>
                     <strong>{{ $componentGroup->name }}</strong>
                     <div class="pull-right text-muted small">
-                        <a href="javascript: void(0);" class="select-group" id="select-all-{{$componentGroup->id}}">Select All</a>
+                        <a href="javascript: void(0);" class="select-group" id="select-all-{{$componentGroup->id}}">{{ trans('cachet.components.select_all') }}</a>
                         &nbsp;|&nbsp;
-                        <a href="javascript: void(0);" class="deselect-group" id="deselect-all-{{$componentGroup->id}}">Deselect All</a>
+                        <a href="javascript: void(0);" class="deselect-group" id="deselect-all-{{$componentGroup->id}}">{{ trans('cachet.components.deselect_all') }}</a>
                     </div>
                 </div>
                 @foreach($componentGroup->enabled_components()->orderBy('order')->get() as $component)
@@ -53,7 +53,7 @@
             @endif
 
             <div class="text-right">
-                <button type="submit" class="btn btn-success">Update Subscription</button>
+                <button type="submit" class="btn btn-success">{{ trans('cachet.subscriber.manage.update_subscription') }}</button>
             </div>
         </form>
     </div>
