@@ -84,6 +84,21 @@ class ComponentTest extends AbstractApiTestCase
         $response->assertJsonFragment(['name' => 'Foo']);
     }
 
+    public function test_can_create_minimal_component()
+    {
+        $this->beUser();
+
+        $this->expectsEvents(ComponentWasCreatedEvent::class);
+
+        $response = $this->json('POST', '/api/v1/components', [
+            'name'        => 'Foo',
+            'status'      => 1,
+        ]);
+
+        $response->assertStatus(200);
+        $response->assertJsonFragment(['name' => 'Foo']);
+    }
+
     public function test_can_create_component_without_enabled_field()
     {
         $this->beUser();
@@ -174,7 +189,7 @@ class ComponentTest extends AbstractApiTestCase
         ]);
 
         $response->assertStatus(200);
-        $response->assertJsonFragment(['name' => 'Foo']);
+        $response->assertJsonFragment(['name' => 'Foo', 'enabled' => $component->enabled]);
     }
 
     public function test_can_update_component_without_status_change()
@@ -190,7 +205,7 @@ class ComponentTest extends AbstractApiTestCase
         ]);
 
         $response->assertStatus(200);
-        $response->assertJsonFragment(['name' => 'Foo']);
+        $response->assertJsonFragment(['name' => 'Foo', 'enabled' => $component->enabled]);
     }
 
     public function test_can_update_component_with_status_change()
@@ -211,7 +226,7 @@ class ComponentTest extends AbstractApiTestCase
         ]);
 
         $response->assertStatus(200);
-        $response->assertJsonFragment(['name' => 'Foo', 'status' => 2]);
+        $response->assertJsonFragment(['name' => 'Foo', 'status' => 2, 'enabled' => $component->enabled]);
     }
 
     public function test_can_update_component_with_meta_data()
@@ -238,6 +253,7 @@ class ComponentTest extends AbstractApiTestCase
                 'uuid' => '172ff3fb-41f7-49d3-8bcd-f57b53627fa0',
                 'foo'  => 'bar',
             ],
+            'enabled' => $component->enabled,
         ]);
     }
 
